+++ release-notes.md
@@ -8,6 +8,7 @@ for [Chrome extension to for Bible.com](README.md).
 
 - [x] Possibility to Display on **2 different screens** (primary => Screen 1, parallel => Screen 2)
 - [x] Export Selected Slide (right click on slide)
+- [x] Allow to upload multiple images for Backgrounds
 
 ## 2.4.0